+++ src/models/title_group_comment.rs
@@ -8,9 +8,9 @@ pub struct TitleGroupComment {
     pub content: String,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
-    pub created_by_id: i32,
+    pub created_by_id: i64,
     pub title_group_id: i64,
-    pub refers_to_torrent_id: Option<i32>,
+    pub refers_to_torrent_id: Option<i64>,
     pub answers_to_comment_id: Option<i64>,
 }
 
@@ -18,6 +18,6 @@ pub struct TitleGroupComment {
 pub struct UserCreatedTitleGroupComment {
     pub content: String,
     pub title_group_id: i64,
-    pub refers_to_torrent_id: Option<i32>,
+    pub refers_to_torrent_id: Option<i64>,
     pub answers_to_comment_id: Option<i64>,
 }
